Add DryRun support for operations

+++ src/S3Batcher/Operations/DeleteVersions.cs
@@ -23,7 +23,7 @@ namespace S3Batcher.Operations
 
         public void Execute(OperationOptions options)
         {
-            _logger.Info($"Starting versions deletion...");
+            _logger.Info($"Starting versions deletion {(options.DryRun ? "(dry run)" : string.Empty)}...");
             var listRequest = new ListVersionsRequest
             {
                 BucketName = options.BucketName,
@@ -43,7 +43,11 @@ namespace S3Batcher.Operations
                 }).ToList()
             };
 
-            _s3Client.DeleteObjectsAsync(deleteRequest).Wait();
+            if (!options.DryRun)
+            {
+                _s3Client.DeleteObjectsAsync(deleteRequest).Wait();
+            }
+
             _logger.Info("Deletion completed!");
         }
     }
